requirements.txt
Django==5.0.7
django_phonenumber_field==8.0.0
geopy==2.4.1
google_api_python_client==2.137.0
pandas==2.2.2
protobuf
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.1
google.generativeai
phonenumbers
